Moved some attribute classes to common directory
