+++ lib/servactory/configuration/setup.rb
@@ -22,7 +22,8 @@ module Servactory
         @collection_mode_class_names =
           Servactory::Maintenance::CollectionMode::ClassNamesCollection.new(default_collection_mode_class_names)
 
-        @input_option_helpers = Servactory::Inputs::OptionHelpersCollection.new(default_input_option_helpers)
+        @input_option_helpers =
+          Servactory::Maintenance::Attributes::OptionHelpersCollection.new(default_input_option_helpers)
 
         @aliases_for_make = Servactory::Methods::AliasesForMake::Collection.new
         @shortcuts_for_make = Servactory::Methods::ShortcutsForMake::Collection.new
@@ -36,7 +37,7 @@ module Servactory
 
       def default_input_option_helpers
         Set[
-          Servactory::Inputs::OptionHelper.new(name: :optional, equivalent: { required: false }),
+          Servactory::Maintenance::Attributes::OptionHelper.new(name: :optional, equivalent: { required: false }),
         ]
       end
     end
