--- lib/servactory/inputs/option_helper.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-module Servactory
-  module Inputs
-    class OptionHelper
-      attr_reader :name,
-                  :equivalent
-
-      def initialize(name:, equivalent:)
-        @name = name
-        @equivalent = equivalent
-      end
-    end
-  end
-end
