--- sig/lib/servactory/inputs/option_helper.rbs
@@ -1,11 +1,0 @@
-module Servactory
-  module Inputs
-    class OptionHelper
-      attr_reader name: Symbol
-
-      attr_reader equivalent: Hash[Symbol, untyped]
-
-      def initialize: (name: Symbol, equivalent: Hash[Symbol, untyped]) -> void
-    end
-  end
-end
