--- sig/lib/servactory/inputs/option_helpers_collection.rbs
@@ -1,13 +1,0 @@
-module Servactory
-  module Inputs
-    class OptionHelpersCollection
-      extend Forwardable
-
-      @collection: Set[OptionHelper]
-
-      def initialize: (Set[OptionHelper] collection) -> void
-
-      def find_by: (name: Symbol) -> OptionHelper?
-    end
-  end
-end
