+++ sig/lib/servactory/maintenance/attributes/option_helper.rbs
@@ -1,0 +1,13 @@
+module Servactory
+  module Maintenance
+    module Attributes
+      class OptionHelper
+        attr_reader name: Symbol
+
+        attr_reader equivalent: Hash[Symbol, untyped]
+
+        def initialize: (name: Symbol, equivalent: Hash[Symbol, untyped]) -> void
+      end
+    end
+  end
+end
